backend/requirements.txt
# 차량 감지 및 세그멘테이션 프로젝트 의존성 패키지
# 버전은 py310_pt 기준으로 고정. ~=는 마이너 업데이트 허용, 메이저는 차단.

# 핵심 패키지
numpy~=1.24.3
opencv-python~=4.11.0
matplotlib~=3.7.2
pillow~=11.2.1

# 딥러닝 프레임워크
torch~=2.4.1
torchvision~=0.19.1

# YOLO 객체 감지
ultralytics~=8.4.14

# 웹 애플리케이션 (레거시 Streamlit, MVP 기간 병행)
streamlit~=1.48.1
pandas~=2.0.3

# REST API
fastapi~=0.116.1
uvicorn[standard]~=0.35.0
pydantic~=2.12.5
python-multipart~=0.0.20
structlog~=24.1

# SAM 세그멘테이션 (선택사항 - 수동 설치 필요)
# segment-anything @ git+https://github.com/facebookresearch/segment-anything.git

# 유틸리티
tqdm~=4.65.0
requests~=2.28.2
psutil~=5.9.8  # 메모리 모니터링용

# 개발 도구
pytest~=7.4
httpx~=0.27.2
black~=23.0
flake8~=6.0

# 주피터 노트북 지원 (선택사항)
# jupyter>=1.0.0
# ipywidgets>=7.6.0
